prints the size of the various type

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int b;
 long int c;
 long long int d;
 float e;
-printf("Size of a char: %li byte(s) \n", sizeof(a));
-printf("Size of a int: %li byte(s)\n", sizeof(b));
-printf("Size of a long int: %li byte(s)\n", sizeof(c));
-printf("Size of a long long int: %li byte(s)\n", sizeof(d));
-printf("Size of a float: %li byte(s)\n", sizeof(e));
+printf("Size of a char: %li byte(s) \n", (unsigned long)sizeof(a));
+printf("Size of a int: %li byte(s)\n", (unsigned long)sizeof(b));
+printf("Size of a long int: %li byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of a long long int: %li byte(s)\n", (unsigned long)sizeof(d));
+printf("Size of a float: %li byte(s)\n", (unsigned long)sizeof(e));
 return (0);
 }
